Added translate to Italian feature in front end
Added button to translate that takes text from textarea and
sends it to the API where it is translated.  The docs service
contains the translate method, which makes a patch call to /translate
and then a get request for the doc.  Once the doc is retrieved, it
dynamically updates the textare text with the new text